pigsheadbbq/server/requirements.txt
flask>=3.0,<4.0
gunicorn>=21.0,<22.0
Flask-WTF>=1.2,<2.0
argon2-cffi>=23.1,<24.0
